Removed render option, added natural as new default
After some research I discovered that waiting for onload to fire on a
detached new Image() object doesn't guarantee complete rendering of an
<img> element with the same source. You'll notice this when loading very
large images.

Having an option called `render` that uses this approach but doesn't
guarantee a full render all the time doesn't make sense, so we're
switching to a new option called `natural`, called as soon as
naturalWidth/Height are available. This explains the intent of the
polling method much better. It's enabled as our new default to make
things faster.